pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["SegmentEditorAdaptiveBrush*"]
exclude = ["SegmentEditorAdaptiveBrushTester*"]

[project]
name = "SlicerAdaptiveBrush"
version = "0.1.0"
description = "Adaptive brush segment editor effect for 3D Slicer"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "SlicerAdaptiveBrush Contributors"}
]
keywords = ["3d-slicer", "medical-imaging", "segmentation", "adaptive-brush"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Processing",
]

[project.urls]
Homepage = "https://github.com/benzwick/SlicerAdaptiveBrush"
Documentation = "https://github.com/benzwick/SlicerAdaptiveBrush#readme"
Repository = "https://github.com/benzwick/SlicerAdaptiveBrush.git"
Issues = "https://github.com/benzwick/SlicerAdaptiveBrush/issues"

# Development dependencies for local dev with uv
# Usage: uv sync --extra dev
[project.optional-dependencies]
dev = [
    "numpy>=1.20.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
    "pre-commit>=4.0.0",
    "optuna>=4.0.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["SegmentEditorAdaptiveBrush/Testing/Python"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["SegmentEditorAdaptiveBrush/SegmentEditorAdaptiveBrushLib"]
omit = ["*/Testing/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["Testing/", "build/", "TestCases/"]

# Ruff replaces black, flake8, isort
[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [".git", "__pycache__", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
]

[tool.ruff.lint.isort]
known-first-party = ["SegmentEditorAdaptiveBrush", "SegmentEditorAdaptiveBrushLib"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
